requirements.txt
# Result of `pip freeze` after installing `requirements.txt.in`
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
discord-py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
typing-extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
